Alter BIST tests stop emitting warnings during runtests invocation.
* hsv2rgb.m, rgb2hsv.m, rgb2ntsc.m: Use intended variable of 'int8' rather
than 'uint16'.

* strread.m: Change %!test to %!warning to verify, but swallow, emitted
warning.  Wrap test to < 80 characters.  Place comments about bugs
above %!test block.  Use %!assert blocks rather than %!test, followed
by assert.  Add missing semicolons.

* ode23.m: Clean up file to follow Octave guidelines for indentation.
Use '#' for comments which follow code on a line.
Use 'warning ("off", ..., "local")' to temporarily disable warning messages
during BIST testing.

* ode45.m: Add %!error tests at end of file.
Use 'warning ("off", ..., "local")' to temporarily disable warning messages
during BIST testing.

* pcg.m: Change %!test to %!warning to verify, but swallow, emitted warning.

* hankel.m: Change %!test to %!warning to verify, but swallow, emitted warning.

+++ scripts/image/hsv2rgb.m
@@ -182,7 +182,7 @@ endfunction
 %! assert (size (rgb), [10 10 3])
 
 %!test
-%! rgb = hsv2rgb (randi ([-128 127], 10, 10, 3, "uint16"));
+%! rgb = hsv2rgb (randi ([-128 127], 10, 10, 3, "int8"));
 %! assert (class (rgb), "double")
 %! assert (size (rgb), [10 10 3])
